Finish landing page; Refactor refresh token logic code; Implement authentication with Google; Add favicon

+++ backend/src/main.py
@@ -1,16 +1,18 @@
-from .schemas import UserSchema, TokenSchema
+from .schemas import UserSchema, TokenSchema, CodeSchema
 from .database import get_session
 from .models import Users
 from .services.users import create_user, get_user, days, calculate_streak
 from .services.tasks import get_uncompleted_tasks, complete, create_habit, create_habits_auto, get_uncompleted_habits, remove_habit
 from .services.learnings import get_learning_day, learning_complete, get_learning_short
-from .services.companions import get_companion, get_accessories, add_accessory, change_companion_type, change_companion_name, update_accessory_visibility
+from .services.companions import get_accessories, add_accessory, change_companion_type, change_companion_name, update_accessory_visibility
 from .auth import utils as au
+from .auth.oauth_google import generate_google_auth_redirect_uri, handle_code
 from .exceptions import DuplicateError, TimeGapError
 from .dependecies import get_current_user
 from sqlalchemy.orm import Session
 from uuid import UUID
-from fastapi import FastAPI, Depends, HTTPException, Response, Cookie, status
+from fastapi import FastAPI, Depends, HTTPException, Response, Cookie, status, Body
+from fastapi.responses import RedirectResponse
 from fastapi.middleware.cors import CORSMiddleware
 from typing import Annotated
 
@@ -40,9 +42,15 @@ def root():
     return {"msg": "Helfy backend is running"}
 
 @app.post("/auth/sign-up")
-def register(user_data: UserSchema, db: Annotated[Session, Depends(get_session)]):
+def register(user_data: UserSchema, db: Annotated[Session, Depends(get_session)], response: Response):
     if create_user(db, user_data):
-        return {"ok": True, "msg": "User has been added successfully"}
+        user = get_user(db, email=user_data.email)
+        access_token = au.encode_jwt(payload={
+            "sub": user.id
+        })
+        refresh_token = au.create_refresh_token(db, user.id)
+        response.set_cookie(key="refresh_token", value=refresh_token, httponly=True, secure=False, samesite="lax", max_age=30*24*60*60)
+        return {"ok": True, "msg": "User has been added successfully", "token": TokenSchema(access_token=access_token, token_type="bearer")}
     else:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User with this email already exsists")
 
@@ -51,6 +59,8 @@ def login(response: Response,
           user_data: UserSchema, 
           db: Annotated[Session, Depends(get_session)]):
     user = get_user(db, email=user_data.email)
+    if user.auth_provider != "local":
+        raise HTTPException(status_code=401, detail="Use Google login for this account")
     if user and au.validate_password(user.password, user_data.password):
         access_token = au.encode_jwt(payload={
             "sub": user.id
@@ -62,32 +72,66 @@ def login(response: Response,
 
 @app.post("/auth/token/refresh")
 def refresh_access_token(db: Annotated[Session, Depends(get_session)], 
-                         response: Response, 
-                         refresh_token: UUID = Cookie(None)):
-    print(f"Here is the token sent by client {refresh_token}")
-    token_result = au.is_valid_refresh_token(db, refresh_token)
-    if token_result == False:
+                         response: Response,
+                         refresh_token: str = Cookie(None)):
+    if not refresh_token:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Missing refresh token")
+    try:
+        user_id = au.is_valid_refresh_token(db, refresh_token)
+    except HTTPException:
         response.delete_cookie("refresh_token")
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid refresh token")
+        raise
     
-    new_refresh_token = au.create_refresh_token(db, user_id=token_result)
+    new_refresh_token = au.create_refresh_token(db, user_id=user_id)
     access_token = au.encode_jwt(payload={
-        "sub": token_result
+        "sub": user_id
     })
-    response.set_cookie(key="refresh_token", value=new_refresh_token, httponly=True, secure=False, samesite="lax", max_age=30*24*60*60)
+    response.set_cookie(key="refresh_token", 
+                        value=new_refresh_token, 
+                        httponly=True, 
+                        secure=False, 
+                        samesite="lax", 
+                        max_age=30*24*60*60)
     return {"ok": True, "token": TokenSchema(access_token=access_token, token_type="bearer")}
 
 @app.post("/auth/log-out")
 def logout(db: Annotated[Session, Depends(get_session)], 
                          response: Response, 
-                         refresh_token: UUID = Cookie(None)):
-    token_result = au.is_valid_refresh_token(db, refresh_token)
-    if token_result == False:
+                         refresh_token: str = Cookie(None)):
+    if not refresh_token:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Missing refresh token")
+    try:
+        token_result = au.is_valid_refresh_token(db, refresh_token)
+    except HTTPException:
         response.delete_cookie("refresh_token")
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid refresh token")
+        raise
     
     response.delete_cookie("refresh_token")
     return {"ok": True, "msg": "Successfully logged out"}
+
+@app.get("/auth/google/url")
+def get_google_oauth_redirect():
+    uri = generate_google_auth_redirect_uri()
+    return RedirectResponse(url=uri, status_code=302)
+
+@app.post("/auth/google/callback")
+def google_callback(db: Annotated[Session, Depends(get_session)], code: CodeSchema, response: Response):
+    email = handle_code(code.code, key="email")
+    if not email:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Unable to retrieve email from Google")
+    user = get_user(db, email=email)
+    if not user:
+        new_user = UserSchema(email=email, password="random", auth_provider="google")
+        create_user(db, new_user)
+        user = get_user(db, email=email)
+    access_token = au.encode_jwt(payload={
+            "sub": user.id
+        })
+    refresh_token = au.create_refresh_token(db, user.id)
+    response.set_cookie(key="refresh_token", 
+                        value=refresh_token, 
+                        httponly=True, secure=False, samesite="lax", max_age=30*24*60*60)
+    return {"ok": True, "token": TokenSchema(access_token=access_token, token_type="bearer")}
 
 @app.get("/users/stats/streak")
 def get_streak(db: Annotated[Session, Depends(get_session)],
@@ -177,10 +221,7 @@ def complete_learning(db: Annotated[Session, Depends(get_session)],
 
 @app.get("/companion")
 def companion(user: Annotated[Users, Depends(get_current_user)]):
-    result = get_companion(user)
-    if not result:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Companion doesn't exist")
-    return {"ok": True, "companion": result}
+    return {"ok": True, "companion": user.companion}
 
 @app.get("/companion/accessories")
 def accessories(db: Annotated[Session, Depends(get_session)],
